project 3 and exc 2
I deleted datasources which I added manually before. it was not allowed and I recently found out.
second change is related to adding pipeline.py and pipeline.sh which is related to project work 3. it is a first draft to get data and save it and not so much change in dataset.
I also updated project-plan due to the change of dataset url.
last change is adding exercise 2 to the git which is completed